+++ training/hyper/hyper_freesolv.py
@@ -978,7 +978,7 @@ hyper = {
                 "inputs": [
                     {"shape": (None,), "name": "node_number", "dtype": "float32", "ragged": True},
                     {"shape": (None, 3), "name": "node_coordinates", "dtype": "float32", "ragged": True},
-                    {"shape": (None, 1), "name": "edge_weights", "dtype": "float32", "ragged": True},  # or edge_weights
+                    {"shape": (None, 1), "name": "edge_weights", "dtype": "float32", "ragged": True},
                     {"shape": (None, 2), "name": "edge_indices", "dtype": "int64", "ragged": True}
                 ],
                 "input_embedding": {"node": {"input_dim": 95, "output_dim": 64},
@@ -986,9 +986,11 @@ hyper = {
                 "use_edge_embedding": False,
                 "max_atoms": None,
                 "distance_matrix_kwargs": {"trafo": "exp"},
-                "attention_kwargs": {"units": 64, "lambda_a": 1.0, "lambda_g": 0.5, "lambda_d": 0.5},
-                "feed_forward_kwargs": {"units": [64, 64, 64], "activation": ["relu", "relu", "linear"]},
-                "embedding_units": 64,
+                "attention_kwargs": {"units": 8, "lambda_attention": 0.3, "lambda_distance": 0.3,
+                                     "lambda_adjacency": None,
+                                     "dropout": 0.1},
+                "feed_forward_kwargs": {"units": [32, 32, 32], "activation": ["relu", "relu", "linear"]},
+                "embedding_units": 32,
                 "depth": 5,
                 "heads": 8,
                 "merge_heads": "concat",
@@ -1031,7 +1033,9 @@ hyper = {
                 "config": {},
                 "methods": [
                     {"set_attributes": {}},
-                    {"map_list": {"method": "normalize_edge_weights_sym"}}
+                    {"map_list": {"method": "normalize_edge_weights_sym"}},
+                    {"map_list": {"method": "pad_property", "key": "node_number", "pad_width": [0, 1]}},
+                    {"map_list": {"method": "pad_property", "key": "node_coordinates", "pad_width": [[0, 1], [0, 0]]}}
                 ]
             },
             "data_unit": "mol/L"