+++ changelog.md
@@ -44,7 +44,7 @@ v2.0.3
 * Improved docs overall.
 * Added ragged node/edge embedding output for TF > 2.8 via "output_to_tensor" model config.
 * Added ``make_function`` option to training scripts.
-* Refactored GraphDict methods into ``kgcnn.data.adapter.GraphMethodsAdapter``.
+* Refactored GraphDict methods into ``kgcnn.data.adapter.GraphTensorMethodsAdapter``.
 * Removed ``kgcnn.layers.modules.ReduceSum`` as it has not been used and may be problematic.
 * Moved ``kgcnn.utils.data`` to ``kgcnn.data.utils``. 
 * Refactored smile to mol generation into ``kgcnn.mol.convert`` and renamed `kgcnn.mol.gen` to `kgcnn.mol.external`